chore(front): variabilize angular dev server port
To allow for several instances of the dev environnment to run on the
same host at the same time (e.g. several worktrees).

Gitignore .env too as it should contain only local overrides of env
vars when they need to be different from default values (e.g. the new
var introduced which need to have a different value in each worktree)